test: require_build_type cleanup
Some tests require to be executed on debug build only, i.e. because they
expect debug traces to be logged.  In practice, non-debug version of
such tests is compiled with the same flags as debug build ("-DDEBUG"),
so the same test is executed twice.
Also, there are tests that always link statically to NVM libraries, so
there is no point in building/executing their static variants.

This patch explicitly disables compilation of static or nondebug
versions of selected tests, as well as adds "require_build_type debug"
to TEST* scripts.